Add tests for auth handler break unit tests for handlers into a separate file Misc error handling fixes in Login handler
