+++ tests/unit/netapi/rest_tornado/test_utils.py
@@ -78,150 +78,7 @@ class TestEventListener(tornado.testing.AsyncTestCase):
             assert event_future.result()['tag'] ==  'evt1'
             assert event_future.result()['data']['data'] ==  'foo1'
 
-class TestBaseSaltAPIHandler(tornado.testing.AsyncHTTPTestCase):
-    content_type_map = {'json': 'application/json',
-                        'yaml': 'application/x-yaml',
-                        'text': 'text/plain',
-                        'form': 'application/x-www-form-urlencoded'}
-    def get_app(self):
-        class StubHandler(saltnado.BaseSaltAPIHandler):
-            def get(self):
-                return self.echo_stuff()
 
-            def post(self):
-                return self.echo_stuff()
-
-            def echo_stuff(self):
-                ret_dict = {'foo': 'bar'}
-                attrs = ('token',
-                         'start',
-                         'connected',
-                         'lowstate',
-                         )
-                for attr in attrs:
-                    ret_dict[attr] = getattr(self, attr)
-
-                self.write(self.serialize(ret_dict))
-
-        return tornado.web.Application([('/', StubHandler)], debug=True)
-
-    def test_content_type(self):
-        '''
-        Test the base handler's accept picking
-        '''
-
-        # send NO accept header, should come back with json
-        response = self.fetch('/')
-        assert response.headers['Content-Type'] == self.content_type_map['json']
-        assert type(json.loads(response.body)) == dict
-
-        # send application/json
-        response = self.fetch('/', headers={'Accept': self.content_type_map['json']})
-        assert response.headers['Content-Type'] == self.content_type_map['json']
-        assert type(json.loads(response.body)) == dict
-
-        # send application/x-yaml
-        response = self.fetch('/', headers={'Accept': self.content_type_map['yaml']})
-        assert response.headers['Content-Type'] == self.content_type_map['yaml']
-        assert type(yaml.load(response.body)) == dict
-
-    def test_token(self):
-        '''
-        Test that the token is returned correctly
-        '''
-        token = json.loads(self.fetch('/').body)['token']
-        assert token is None
-
-        # send a token as a header
-        response = self.fetch('/', headers={saltnado.AUTH_TOKEN_HEADER: 'foo'})
-        token = json.loads(response.body)['token']
-        assert token == 'foo'
-
-        # send a token as a cookie
-        response = self.fetch('/', headers={'Cookie': '{0}=foo'.format(saltnado.AUTH_COOKIE_NAME)})
-        token = json.loads(response.body)['token']
-        assert token == 'foo'
-
-        # send both, make sure its the header
-        response = self.fetch('/', headers={saltnado.AUTH_TOKEN_HEADER: 'foo',
-                                            'Cookie': '{0}=bar'.format(saltnado.AUTH_COOKIE_NAME)})
-        token = json.loads(response.body)['token']
-        assert token == 'foo'
-
-    # TODO: break into separate functions per type?
-    def test_deserialize(self):
-        '''
-        Send various encoded forms of lowstates (and bad ones) to make sure we
-        handle deserialization correctly
-        '''
-        valid_lowstate = [{
-                "client": "local",
-                "tgt": "*",
-                "fun": "test.fib",
-                "arg": ["10"]
-            },
-            {
-                "client": "runner",
-                "fun": "jobs.lookup_jid",
-                "jid": "20130603122505459265"
-            }]
-
-        # send as JSON
-        response = self.fetch('/',
-                              method='POST',
-                              body=json.dumps(valid_lowstate),
-                              headers={'Content-Type': self.content_type_map['json']})
-
-        assert valid_lowstate == json.loads(response.body)['lowstate']
-
-        # send yaml as json (should break)
-        response = self.fetch('/',
-                              method='POST',
-                              body=yaml.dump(valid_lowstate),
-                              headers={'Content-Type': self.content_type_map['json']})
-        assert response.code == 400
-
-        # send as yaml
-        response = self.fetch('/',
-                              method='POST',
-                              body=yaml.dump(valid_lowstate),
-                              headers={'Content-Type': self.content_type_map['yaml']})
-        assert valid_lowstate == json.loads(response.body)['lowstate']
-
-        # send json as yaml (works since yaml is a superset of json)
-        response = self.fetch('/',
-                              method='POST',
-                              body=json.dumps(valid_lowstate),
-                              headers={'Content-Type': self.content_type_map['yaml']})
-        assert valid_lowstate == json.loads(response.body)['lowstate']
-
-        # send json as text/plain
-        response = self.fetch('/',
-                              method='POST',
-                              body=json.dumps(valid_lowstate),
-                              headers={'Content-Type': self.content_type_map['text']})
-        assert valid_lowstate == json.loads(response.body)['lowstate']
-
-
-        # send form-urlencoded
-        form_lowstate = (
-            ('client', 'local'),
-            ('tgt', '*'),
-            ('fun', 'test.fib'),
-            ('arg', '10'),
-        )
-        response = self.fetch('/',
-                              method='POST',
-                              body=urllib.urlencode(form_lowstate),
-                              headers={'Content-Type': self.content_type_map['form']})
-        returned_lowstate = json.loads(response.body)['lowstate']
-        assert len(returned_lowstate) == 1
-        returned_lowstate = returned_lowstate[0]
-
-        assert returned_lowstate['client'] == 'local'
-        assert returned_lowstate['tgt'] == '*'
-        assert returned_lowstate['fun'] == 'test.fib'
-        assert returned_lowstate['arg'] == ['10']
